Fixed all issues and added eslint
