+++ frontend/README.md
@@ -4,6 +4,34 @@ This project was bootstrapped with [Create React App](https://github.com/faceboo
 
 You should not need to run it directly from this folder as it is part of the docker build file but if you want to run it locally for development you can follow the instructions below at "Available Scripts".
 
+## Web App Design
+
+The web app is using a flavour of the MVVM (Model-View-View Model) pattern. You should be aware of the following ideas:
+
+### Components
+
+Here we have simple React code combined with imports of CSS files for the formatting. These components do not have state and you could inject anything you like through the props to test them.
+
+### Controllers
+
+A controller wraps a Component and maps the functions and values coming from a View Model (see below). You might have the occasional useState for maybe an [open, setOpen] value but nothing more.
+
+### State
+
+For state Recoil is being used but the state interacts with the ViewModel and should not be directly accessed or mutated from the Controllers with the exception of the App.tsx file that sets the setters and getters in the View Models.
+
+### View Models
+
+A View Model is a class that has a clearly defined interface that provides controllers with all the necessary functions and values to feed to their Components. We are utilizing the concept of CQRS meaning that we have separate functions that cause side-effects (aka mutations) and separate values that work like queries.
+
+### Repositories
+
+Repositories are used to interact with the app state (e.g. localStorage) and the Services (see below). For example, a service expects some authentication metadata (JWT) with the requests and instead of complicating the ViewModel with these details, a Repository provides a cleaner interface to the ViewModel for using the Services by taking care of the JWT injection etc. You can also use Repositories to deal with local caching etc.
+
+### Services
+
+Services are the gRPC endpoints that are automatically generated by the protobuf file and you could also add other services to wrap 3rd party API calls (e.g. Google Maps API).
+
 ## Technologies Used
 
 To communicate with the backend, grpc-web is being used for the main functionality while a couple REST requests are used for registration and login.
