+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "todo-frontend",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
     "@chakra-ui/icons": "^2.1.0",
@@ -8,23 +8,12 @@
     "@chakra-ui/system": "^2.6.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@grpc/grpc-js": "^1.8.21",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
-    "axios": "^1.3.5",
-    "eslint": "^8.45.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.28.0",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.33.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "@grpc/grpc-js": "^1.8.21",    
     "eventemitter3": "^5.0.1",
     "framer-motion": "^10.15.0",
     "google-protobuf": "^3.21.2",
     "grpc-web": "^1.4.2",
     "jwt-decode": "^3.1.2",
-    "mobx": "^6.10.0",
-    "mobx-react-lite": "^4.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
@@ -37,11 +26,12 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "proto": "protoc -I=. src/bitloops/proto/todo.proto  --js_out=import_style=commonjs,binary:.  --grpc-web_out=import_style=typescript,mode=grpcwebtext:.",
-    "newproto": "protoc -I=. src/bitloops/proto/todo.proto --ts_out=. --ts_opt=target=web,json_names,unary_rpc_promise=true,no_namespace --grpc-web_out=import_style=typescript,mode=grpcwebtext:.",
+    "proto": "protoc -I=. src/bitloops/proto/todo.proto --ts_out=. --ts_opt=target=web,json_names,unary_rpc_promise=true,no_namespace --grpc-web_out=import_style=typescript,mode=grpcwebtext:.",
     "docker:build": "docker build -t todo-frontend .",
     "docker:run": "docker run -dp 3000:3000 --name todo-frontend todo-frontend",
-    "docker": "docker build -t todo-frontend . && docker run -dp 3000:3000 --name todo-frontend todo-frontend"
+    "docker": "docker build -t todo-frontend . && docker run -dp 3000:3000 --name todo-frontend todo-frontend",
+    "postinstall": "patch-package",
+    "lint": "eslint ."
   },
   "eslintConfig": {
     "extends": [
@@ -62,21 +52,37 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/google-protobuf": "^3.15.6",
-    "@types/jest": "^29.5.0",
-    "@types/node": "^20.4.5",
-    "@types/react": "^18.0.34",
-    "@types/react-dom": "^18.0.11",
-    "eslint-config-prettier": "^8.8.0",
+    "@typescript-eslint/eslint-plugin": ">=6.0.0",
+    "@typescript-eslint/parser": ">=6.0.0",
+    "eslint": ">=8.0.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^8.9.0",
+    "eslint-config-xo": "^0.43.1",
+    "eslint-config-xo-typescript": "^1.0.1",
+    "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "prettier": "^3.0.0",
-    "protoc-gen-grpc-web": "^1.4.1",
+    
     "react-scripts": "5.0.1",
-    "typescript": "^5.0.4",
-    "web-vitals": "^3.3.1"
+    "typescript": ">=4.7",
+    "web-vitals": "^2.1.0",
+    
+    "@types/google-protobuf": "^3.15.6",
+    "patch-package": "^8.0.0",
+    "postinstall-postinstall": "^2.1.0",
+    "protoc-gen-grpc-web": "^1.4.2"
   }
 }
