+++ frontend/patches/grpc-web+1.4.2.patch
@@ -1,0 +1,15 @@
+diff --git a/node_modules/grpc-web/index.d.ts b/node_modules/grpc-web/index.d.ts
+index 09fb671..9b6b233 100644
+--- a/node_modules/grpc-web/index.d.ts
++++ b/node_modules/grpc-web/index.d.ts
+@@ -71,8 +71,8 @@ declare module "grpc-web" {
+   export class MethodDescriptor<REQ, RESP> {
+     constructor(name: string,
+                 methodType: string,
+-                requestType: new (...args: unknown[]) => REQ,
+-                responseType: new (...args: unknown[]) => RESP,
++                requestType: new (...args: any) => REQ,
++                responseType: new (...args: any) => RESP,
+                 requestSerializeFn: any,
+                 responseDeserializeFn: any);
+     getName(): string;
