+++ frontend/src/infra/repositories/iam/index.ts
@@ -1,18 +1,3 @@
-import { User } from '../../../models/User';
+import { IamRepository } from './IamRepository';
 
-export type IamResponse = {
-  status: 'success' | 'cached';
-  user: User;
-};
-
-interface IIamRepository {
-  loginWithEmailPassword(email: string, password: string): Promise<User>;
-  registerWithEmailPassword(email: string, password: string): Promise<void>;
-  isAuthenticated(): boolean;
-  logout(): void;
-  getUser(): User | null;
-  setUser(user: User): void;
-}
-
-export { IamRepository } from './IamRepository';
-export type { IIamRepository };
+export default IamRepository;
