+++ frontend/src/infra/services/IamService.ts
@@ -1,15 +1,35 @@
-import axios from 'axios';
 import { AUTH_URL, REGISTRATION_URL } from '../../config';
 import { IIamService, LoginResponse } from '../interfaces/IIamService';
 
 class IamService implements IIamService {
   async loginWithEmailPassword(email: string, password: string): Promise<LoginResponse> {
-    const response = await axios.post<LoginResponse>(AUTH_URL, { email, password });
-    return response.data;
+    const response = await fetch(AUTH_URL, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ email, password }),
+    });
+
+    if (!response.ok) {
+      throw new Error('Network response was not ok');
+    }
+
+    return (await response.json()) as LoginResponse;
   }
 
   async registerWithEmailPassword(email: string, password: string): Promise<void> {
-    await axios.post(REGISTRATION_URL, { email, password });
+    const response = await fetch(REGISTRATION_URL, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ email, password }),
+    });
+
+    if (!response.ok) {
+      throw new Error('Network response was not ok');
+    }
   }
 }
